data_downloader.py
import os
import sqlite3
import pandas as pd
from binance.client import Client
from datetime import datetime, timedelta
import time
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

class BinanceDataDownloader:
    def __init__(self):
        self.api_key = os.getenv('BINANCE_API_KEY')
        self.api_secret = os.getenv('BINANCE_API_SECRET')
        self.client = Client(self.api_key, self.api_secret)
        self.db_path = os.getenv('DATABASE_PATH', 'bitcoin_data.db')
        
    def create_database(self):
        """Create SQLite database and table for storing OHLCV data"""
        conn = sqlite3.connect(self.db_path)
        cursor = conn.cursor()
        
        # Create table for hourly OHLCV data
        cursor.execute('''
            CREATE TABLE IF NOT EXISTS hourly_data (
                id INTEGER PRIMARY KEY AUTOINCREMENT,
                timestamp INTEGER UNIQUE,
                datetime TEXT,
                open_price REAL,
                high_price REAL,
                low_price REAL,
                close_price REAL,
                volume REAL,
                quote_volume REAL,
                trades_count INTEGER,
                taker_buy_base_volume REAL,
                taker_buy_quote_volume REAL
            )
        ''')
        
        # Create index on timestamp for faster queries
        cursor.execute('CREATE INDEX IF NOT EXISTS idx_timestamp ON hourly_data(timestamp)')
        
        conn.commit()
        conn.close()
        print(f"Database created/verified: {self.db_path}")
    
    def get_existing_data_range(self):
        """Check what data already exists in the database"""
        conn = sqlite3.connect(self.db_path)
        cursor = conn.cursor()
        
        cursor.execute('SELECT MIN(timestamp), MAX(timestamp), COUNT(*) FROM hourly_data')
        result = cursor.fetchone()
        conn.close()
        
        if result[0] is None:
            return None, None, 0
        
        min_ts = result[0]
        max_ts = result[1]
        count = result[2]
        
        min_date = datetime.fromtimestamp(min_ts / 1000)
        max_date = datetime.fromtimestamp(max_ts / 1000)
        
        print(f"Existing data: {count} records from {min_date} to {max_date}")
        return min_ts, max_ts, count
    
    def download_historical_data(self, symbol='BTCUSDT', years=5):
        """Download historical hourly data from Binance"""
        print(f"Starting download of {years} years of hourly data for {symbol}")
        
        # Calculate date range
        end_date = datetime.now()
        start_date = end_date - timedelta(days=years * 365)
        
        print(f"Date range: {start_date} to {end_date}")
        
        # Check existing data
        min_existing, max_existing, existing_count = self.get_existing_data_range()
        
        # Convert to timestamps
        start_ts = int(start_date.timestamp() * 1000)
        end_ts = int(end_date.timestamp() * 1000)
        
        # Determine what data to download
        download_ranges = []
        
        if existing_count == 0:
            # No existing data, download everything
            download_ranges.append((start_ts, end_ts))
            print("No existing data found. Downloading full range.")
        else:
            # Fill gaps in existing data
            if start_ts < min_existing:
                download_ranges.append((start_ts, min_existing))
                print(f"Downloading data before existing range: {start_date} to {datetime.fromtimestamp(min_existing/1000)}")
            
            if max_existing < end_ts:
                download_ranges.append((max_existing, end_ts))
                print(f"Downloading data after existing range: {datetime.fromtimestamp(max_existing/1000)} to {end_date}")
        
        if not download_ranges:
            print("All data already exists in database.")
            return
        
        # Download data in chunks
        conn = sqlite3.connect(self.db_path)
        total_records = 0
        
        for start_chunk, end_chunk in download_ranges:
            current_start = start_chunk
            
            while current_start < end_chunk:
                try:
                    # Binance allows max 1000 klines per request
                    # For hourly data, 1000 hours = ~41.7 days
                    print(f"Downloading from {datetime.fromtimestamp(current_start/1000)}...")
                    
                    klines = self.client.get_historical_klines(
                        symbol=symbol,
                        interval=Client.KLINE_INTERVAL_1HOUR,
                        start_str=str(current_start),
                        end_str=str(end_chunk),
                        limit=1000
                    )
                    
                    if not klines:
                        print("No more data available")
                        break
                    
                    # Process and insert data
                    records_inserted = self.insert_klines_data(conn, klines)
                    total_records += records_inserted
                    
                    print(f"Inserted {records_inserted} records. Total: {total_records}")
                    
                    # Update current_start to the timestamp of the last kline + 1 hour
                    last_timestamp = klines[-1][0]
                    current_start = last_timestamp + 3600000  # Add 1 hour in milliseconds
                    
                    # Rate limiting - Binance allows 1200 requests per minute
                    time.sleep(0.1)  # Small delay to avoid rate limits
                    
                except Exception as e:
                    print(f"Error downloading data: {e}")
                    time.sleep(5)  # Wait before retrying
                    continue
        
        conn.close()
        print(f"Download completed! Total records inserted: {total_records}")
    
    def insert_klines_data(self, conn, klines):
        """Insert klines data into database"""
        cursor = conn.cursor()
        records_inserted = 0
        
        for kline in klines:
            try:
                timestamp = kline[0]
                datetime_str = datetime.fromtimestamp(timestamp / 1000).strftime('%Y-%m-%d %H:%M:%S')
                
                cursor.execute('''
                    INSERT OR IGNORE INTO hourly_data (
                        timestamp, datetime, open_price, high_price, low_price, 
                        close_price, volume, quote_volume, trades_count,
                        taker_buy_base_volume, taker_buy_quote_volume
                    ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
                ''', (
                    timestamp,
                    datetime_str,
                    float(kline[1]),  # open
                    float(kline[2]),  # high
                    float(kline[3]),  # low
                    float(kline[4]),  # close
                    float(kline[5]),  # volume
                    float(kline[7]),  # quote volume
                    int(kline[8]),    # trades count
                    float(kline[9]),  # taker buy base volume
                    float(kline[10])  # taker buy quote volume
                ))
                
                if cursor.rowcount > 0:
                    records_inserted += 1
                    
            except Exception as e:
                print(f"Error inserting record: {e}")
                continue
        
        conn.commit()
        return records_inserted
    
    def verify_data(self):
        """Verify the downloaded data"""
        conn = sqlite3.connect(self.db_path)
        
        # Get basic statistics
        query = '''
            SELECT 
                COUNT(*) as total_records,
                MIN(datetime) as earliest_date,
                MAX(datetime) as latest_date,
                MIN(close_price) as min_price,
                MAX(close_price) as max_price,
                AVG(close_price) as avg_price
            FROM hourly_data
        '''
        
        df = pd.read_sql_query(query, conn)
        conn.close()
        
        print("\n=== Data Verification ===")
        print(f"Total records: {df['total_records'].iloc[0]:,}")
        print(f"Date range: {df['earliest_date'].iloc[0]} to {df['latest_date'].iloc[0]}")
        print(f"Price range: ${df['min_price'].iloc[0]:,.2f} - ${df['max_price'].iloc[0]:,.2f}")
        print(f"Average price: ${df['avg_price'].iloc[0]:,.2f}")
        
        # Check for gaps in data (simplified approach for SQLite compatibility)
        conn = sqlite3.connect(self.db_path)
        
        # Get consecutive timestamps to check for gaps
        gap_query = '''
            SELECT 
                h1.datetime as current_datetime,
                h2.datetime as prev_datetime,
                (h1.timestamp - h2.timestamp) / 3600000.0 as hour_gap
            FROM hourly_data h1
            JOIN hourly_data h2 ON h2.timestamp = (
                SELECT MAX(timestamp) 
                FROM hourly_data 
                WHERE timestamp < h1.timestamp
            )
            WHERE (h1.timestamp - h2.timestamp) > 3600000
            ORDER BY h1.timestamp
            LIMIT 10
        '''
        
        try:
            gaps_df = pd.read_sql_query(gap_query, conn)
            conn.close()
            
            if not gaps_df.empty:
                print(f"\nFound {len(gaps_df)} gaps in data (showing first 10):")
                for _, row in gaps_df.iterrows():
                    print(f"Gap of {row['hour_gap']:.1f} hours between {row['prev_datetime']} and {row['current_datetime']}")
            else:
                print("\nNo significant gaps found in the data.")
        except Exception as e:
            conn.close()
            print(f"\nGap analysis skipped due to SQLite limitations: {e}")
            print("Data appears to be continuous based on record count.")

def main():
    downloader = BinanceDataDownloader()
    
    # Create database
    downloader.create_database()
    
    # Download 5 years of hourly data
    downloader.download_historical_data(symbol='BTCUSDT', years=5)
    
    # Verify the data
    downloader.verify_data()
    
    print("\nData download completed successfully!")
    print(f"Database saved as: {downloader.db_path}")

if __name__ == "__main__":
    main()