requirements.txt
python-binance==1.0.19
pandas==2.1.4
python-dotenv==1.0.0
numpy==1.24.3
requests==2.31.0